requirements.txt
fastapi
uvicorn
aiohttp
python-multipart  
python-dotenv
